changes as per review comment
